Add macro support for custom attributes on containers
